+++ examples/asg_ec2/main.tf
@@ -27,7 +27,7 @@ data "aws_security_group" "default" {
 
 data "aws_ami" "amazon_linux" {
   most_recent = true
-  owners      = ["137112412989"] # Amazon
+  owners      = ["amazon"]
 
   filter {
     name = "name"
@@ -57,6 +57,13 @@ resource "aws_iam_service_linked_role" "autoscaling" {
   }
 }
 
+locals {
+  user_data = <<EOF
+#!/bin/bash
+echo "Hello Terraform!"
+EOF
+}
+
 ######
 # Launch configuration and autoscaling group
 ######
@@ -76,6 +83,8 @@ module "example" {
   security_groups              = [data.aws_security_group.default.id]
   associate_public_ip_address  = true
   recreate_asg_when_lc_changes = true
+
+  user_data_base64 = base64encode(local.user_data)
 
   ebs_block_device = [
     {
@@ -100,7 +109,7 @@ module "example" {
   health_check_type         = "EC2"
   min_size                  = 0
   max_size                  = 1
-  desired_capacity          = 0
+  desired_capacity          = 1
   wait_for_capacity_timeout = 0
   service_linked_role_arn   = aws_iam_service_linked_role.autoscaling.arn
 
@@ -115,6 +124,16 @@ module "example" {
       value               = "megasecret"
       propagate_at_launch = true
     },
+    {
+      key                 = "foo"
+      value               = ""
+      propagate_at_launch = true
+    },
+    {
+      key                 = "bar"
+      value               = ""
+      propagate_at_launch = true
+    }
   ]
 
   tags_as_map = {