+++ examples/asg_elb/versions.tf
@@ -2,8 +2,6 @@ terraform {
   required_version = ">= 0.12.6"
 
   required_providers {
-    aws    = ">= 2.41"
-    random = ">= 2.0"
-    null   = ">= 2.0"
+    aws = ">= 2.41"
   }
 }